+++ src/components/google-analytics/index.ts
@@ -1,0 +1,2 @@
+export * from "./google-analytics"
+export * from "./gtag"
