--- src/components/googleAnalytics/index.ts
@@ -1,1 +1,0 @@
-export * from "./gtag"
