+++ src/firebase/client/init-script.tsx
@@ -1,5 +1,5 @@
 "use client"
-import { event } from "@/components/googleAnalytics/gtag"
+import { event } from "@/components/google-analytics/gtag"
 import {
   getToken,
   initializeAppCheck,
